requirements.txt
async-timeout==4.0.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cvxopt==1.3.0
cycler==0.11.0
decorator==4.4.2
Deprecated==1.2.13
flask-cors==3.0.10
Flask==2.1.2
fonttools==4.33.3
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
moviepy==1.0.3
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.1.1
proglog==0.1.10
protobuf==4.21.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyVimeo==1.1.0
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
robotarium-python-simulator==0.0.0
rq==1.10.1
rsa==4.8
scipy==1.8.1
six==1.16.0
tinydb==4.7.0
tqdm==4.64.0
tuspy==0.2.5
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
